requirements.txt
numpy==1.22.4
PyYAML==6.0.1
torch==1.12.0+cu113
torchvision==0.13.0+cu113
tqdm==4.64.0
